http: add runtime reject_unsupported_transfer_encodings back
It allows "TE: chunked, chunked" and double "TE: chunked" in http
response
It allows and double "TE: chunked" in http request

The risk is that it allows unknown TE as well, such as "TE: gzip,
chunked" by just ignoring it. The behavior can be back tracked to
istio 1.4 or envoy 1.12.

Change-Id: I1ab39fdde2b1acc1dd3e15c07db06456b6066683
Signed-off-by: Yuchen Dai <lambdai@google.com>